fix(coordinator): return outage end time in next_connectivity instead of start time (#108)
When not currently in an outage, next_connectivity was incorrectly
returning the start time of the next outage (event.start) instead of
the end time (event.end). This caused next_connectivity to match
next_planned_outage, which is incorrect behavior.

The sensor should return when connectivity will be restored:
- If currently in outage: return end time of current outage
- If not in outage: return end time of next outage

Fixes #103